pyproject.toml
[tool.poetry]
name = "flask-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
setuptools = "^74.1.1"
psycopg2 = "^2.9.9"
flask = "2.0.2"
werkzeug = "2.0.3"
jinja2 = "3.0.2"
flask-login = "0.5.0"
flask-migrate = "3.1.0"
wtforms = "3.0.0"
flask-wtf = "1.0.0"
flask-sqlalchemy = "2.5.1"
sqlalchemy = "1.4.29"
flask-restx = "0.5.1"
flask-minify = "0.37"
pyjwt = "2.5.0"
flask-mail = "0.9.1"
flask-jwt-extended = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
